+++ CHANGELOG.md
@@ -31,6 +31,7 @@ title: Changelog
 -   TypeDoc now recognizes `txt` as an alias of `text` to indicate a code block should not be highlighted.
 -   Changed default of `suppressCommentWarningsInDeclarationFiles` to `true`
 -   Items which are hidden with `@ignore` or `@hidden` but still referenced by other types will no longer produce warnings about not being exported.
+-   TypeDoc will now use the first signature's comment for later signatures in overloads if present, #2718.
 
 TODO: